requirements.txt
tqdm==4.48.2
tensorflow==2.3.1
torch==1.5.0
scikit_image==0.17.2
Cython==0.29.21
albumentations==0.4.6
cffi==1.14.0
scipy==1.5.2
imageio==2.9.0
numpy==1.16.6
torchvision==0.6.0a0+82fd1c8
matplotlib==3.3.1
h5py==2.10.0
opencv_python==4.4.0.42
easydict==1.9
ipython==7.23.1
Pillow==8.2.0
predict==3.1.13
PyYAML==5.4.1
scikit_learn==0.24.2
skimage==0.0
tensorboardX==2.2
